Split payment-helpers tests into a folder and drop its max-lines exemption
The payment-helpers test file had grown past 1,000 lines and was
grandfathered into biome's `noExcessiveLinesPerFile` exemption list.
Split it into `test/lib/payment-helpers/` — grouped by what each part
checks (metadata build/round-trip, dispatch helpers, provider limits,
and the env-backed buildItemsMetadata flows) — so every file sits well
under the limit, then removed the file's exemption entry.

All original tests carry over unchanged and pass. jscpd stays at 0%
duplication after the split.
