+++ biome.json
@@ -163,7 +163,6 @@
         "test/e2e/accounting.test.ts",
         "test/e2e/duration-days.test.ts",
         "test/lib/attendee-merge.test.ts",
-        "test/lib/payment-helpers.test.ts",
         "test/lib/server-attendee-form.test.ts",
         "test/lib/server-attendees.test.ts",
         "test/lib/server-groups.test.ts",
