+++ test/lib/payment-helpers/dispatch.test.ts
@@ -1,0 +1,173 @@
+import { expect } from "@std/expect";
+import { describe, it as test } from "@std/testing/bdd";
+import { ErrorCode } from "#shared/logger.ts";
+import {
+  createWithClient,
+  errorMessage,
+  hasRequiredSessionMetadata,
+  PaymentUserError,
+  safeAsync,
+  toCheckoutResult,
+} from "#shared/payment-helpers.ts";
+import { isPaymentStatus } from "#shared/payments.ts";
+
+describe("payment-helpers", () => {
+  describe("hasRequiredSessionMetadata", () => {
+    test("returns false for null/undefined", () => {
+      expect(hasRequiredSessionMetadata(null)).toBe(false);
+      expect(hasRequiredSessionMetadata(undefined)).toBe(false);
+    });
+
+    test("returns false when name is missing or empty", () => {
+      expect(
+        hasRequiredSessionMetadata({ email: "a@b.com", items: "[]" }),
+      ).toBe(false);
+      expect(
+        hasRequiredSessionMetadata({
+          email: "a@b.com",
+          items: "[]",
+          name: "",
+        }),
+      ).toBe(false);
+    });
+
+    test("returns false when items missing", () => {
+      expect(
+        hasRequiredSessionMetadata({ email: "a@b.com", name: "Alice" }),
+      ).toBe(false);
+    });
+
+    test("returns true for valid single-listing (email optional)", () => {
+      expect(hasRequiredSessionMetadata({ items: "[]", name: "Alice" })).toBe(
+        true,
+      );
+      expect(
+        hasRequiredSessionMetadata({ email: "", items: "[]", name: "Alice" }),
+      ).toBe(true);
+    });
+
+    test("returns true for valid multi-listing metadata", () => {
+      expect(
+        hasRequiredSessionMetadata({
+          email: "a@b.com",
+          items: '[{"e":1,"q":2,"p":2000}]',
+          name: "Alice",
+        }),
+      ).toBe(true);
+    });
+  });
+
+  describe("isPaymentStatus", () => {
+    test("accepts valid statuses", () => {
+      expect(isPaymentStatus("paid")).toBe(true);
+      expect(isPaymentStatus("unpaid")).toBe(true);
+      expect(isPaymentStatus("no_payment_required")).toBe(true);
+    });
+
+    test("rejects invalid values", () => {
+      expect(isPaymentStatus("completed")).toBe(false);
+      expect(isPaymentStatus("")).toBe(false);
+    });
+  });
+
+  describe("errorMessage", () => {
+    test("extracts message from Error, returns fallback for non-Error", () => {
+      expect(errorMessage(new Error("broke"))).toBe("broke");
+      expect(errorMessage("string")).toBe("Unknown error");
+      expect(errorMessage(null)).toBe("Unknown error");
+    });
+  });
+
+  describe("safeAsync", () => {
+    test("returns value on success, null on error", async () => {
+      expect(
+        await safeAsync(() => Promise.resolve(42), ErrorCode.PAYMENT_CHECKOUT),
+      ).toBe(42);
+      expect(
+        await safeAsync(
+          () => Promise.reject(new Error("boom")),
+          ErrorCode.PAYMENT_CHECKOUT,
+        ),
+      ).toBeNull();
+      expect(
+        await safeAsync(
+          () => Promise.reject("string error"),
+          ErrorCode.PAYMENT_CHECKOUT,
+        ),
+      ).toBeNull();
+    });
+
+    test("re-throws PaymentUserError", async () => {
+      await expect(
+        safeAsync(
+          () => Promise.reject(new PaymentUserError("Bad phone")),
+          ErrorCode.PAYMENT_CHECKOUT,
+        ),
+      ).rejects.toThrow("Bad phone");
+    });
+  });
+
+  describe("createWithClient", () => {
+    test("returns null when client is null", async () => {
+      const withClient = createWithClient(() => Promise.resolve(null));
+      const result = await withClient(
+        () => Promise.resolve("value"),
+        ErrorCode.PAYMENT_CHECKOUT,
+      );
+      expect(result).toBeNull();
+    });
+
+    test("passes client to operation", async () => {
+      const withClient = createWithClient(() =>
+        Promise.resolve({ token: "abc" }),
+      );
+      expect(
+        await withClient(
+          (c) => Promise.resolve(`got-${c.token}`),
+          ErrorCode.PAYMENT_CHECKOUT,
+        ),
+      ).toBe("got-abc");
+    });
+
+    test("returns null on operation error, re-throws PaymentUserError", async () => {
+      const withClient = createWithClient(() =>
+        Promise.resolve({ token: "abc" }),
+      );
+      expect(
+        await withClient(
+          () => Promise.reject(new Error("fail")),
+          ErrorCode.PAYMENT_CHECKOUT,
+        ),
+      ).toBeNull();
+      await expect(
+        withClient(
+          () => Promise.reject(new PaymentUserError("Phone invalid")),
+          ErrorCode.PAYMENT_CHECKOUT,
+        ),
+      ).rejects.toThrow("Phone invalid");
+    });
+  });
+
+  describe("toCheckoutResult", () => {
+    test("returns result when both id and url present", () => {
+      expect(
+        toCheckoutResult("sess_1", "https://pay.example.com", "Stripe"),
+      ).toEqual({
+        checkoutUrl: "https://pay.example.com",
+        sessionId: "sess_1",
+      });
+    });
+
+    test("returns null for missing or empty id/url", () => {
+      expect(
+        toCheckoutResult(undefined, "https://pay.example.com", "Stripe"),
+      ).toBeNull();
+      expect(toCheckoutResult("sess_1", undefined, "Stripe")).toBeNull();
+      expect(toCheckoutResult("sess_1", null, "Stripe")).toBeNull();
+      expect(
+        toCheckoutResult("", "https://pay.example.com", "Stripe"),
+      ).toBeNull();
+      expect(toCheckoutResult("sess_1", "", "Payment")).toBeNull();
+    });
+  });
+});
